pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "seaview"
version = "0.1.0"
description = "Cruise support package for generating oceanographic map tiles"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Bror Jonsson", email = "brorfred@gmail.com"}
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
keywords = ["oceanography", "satellite", "tiles", "maps", "cruise", "copernicus"]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "xarray>=2023.1.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pillow>=9.0.0",
    "mercantile>=1.2.0",
    "netcdf4>=1.6.0",
    "satpy>=0.45.0",
    "pyresample>=1.27.0",
    "pyproj>=3.5.0",
    "copernicusmarine>=2.0.0",
    "sysrsync>=1.1.0",
    "tqdm>=4.65.0",
    #"fsspec>=2023.1.0",
    "dynaconf>=3.2.12",
    "cmasher>=1.9.2",
    "jinja2>=3.1.6",
    "typer>=0.21.1",
    "h5py>=3.15.1",
    "h5netcdf>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
all = [
    "tables>=3.8.0",
    "zarr>=2.14.0",
    "scikit-learn>=1.2.0",
    "eumdac>=2.0.0",
    "rasterio>=1.3.0",
    "rio-tiler>=6.0.0",
    "tinydb>=4.7.0",
    "dynaconf>=3.1.0",
    "folium>=0.14.0",
    "ipython>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/brorfred/seaview"
Documentation = "https://github.com/brorfred/seaview#readme"
Repository = "https://github.com/brorfred/seaview.git"
Issues = "https://github.com/brorfred/seaview/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/seaview"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.uv.sources]
seaview = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ipython>=8.0.0",
    "seaview",
]

[project.scripts]
sea = "seaview.cli2:app"
sea-update = "seaview.cli2:update"
